Sincronizar cambios locales: añadir script importación, config PHP, ajustes Docker y excluir .env de git

+++ database/migrations/2026_02_28_222900_sync_manual_db_changes.php
@@ -30,7 +30,7 @@ return new class extends Migration {
         if (!Schema::hasTable('comentarios')) {
             Schema::create('comentarios', function (Blueprint $table) {
                 $table->id();
-                $table->unsignedInteger('employee_id');
+                $table->unsignedBigInteger('employee_id');
                 $table->text('comment');
                 $table->timestamps();
